+++ src/components/anchor-navigation/anchor-navigation-test.stories.tsx
@@ -4,7 +4,7 @@ import { Meta, StoryObj } from "@storybook/react";
 
 import Textbox from "../textbox";
 import Button from "../button";
-import DialogFullScreen from "../dialog-full-screen";
+import Dialog from "../dialog";
 import {
   AnchorNavigation,
   AnchorSectionDivider,
@@ -99,7 +99,8 @@ export const InFullScreenDialogStory: Story = () => {
   return (
     <>
       <Button onClick={() => setIsOpen(true)}>Open AnchorNavigation</Button>
-      <DialogFullScreen
+      <Dialog
+        fullscreen
         open={isOpen}
         onCancel={() => setIsOpen(false)}
         title="Title"
@@ -138,7 +139,7 @@ export const InFullScreenDialogStory: Story = () => {
             <Content title="Fifth section" />
           </div>
         </AnchorNavigation>
-      </DialogFullScreen>
+      </Dialog>
     </>
   );
 };
