+++ src/components/anchor-navigation/anchor-navigation.mdx
@@ -92,9 +92,9 @@ const MyComponent = () => (
 
 <Story of={AnchorNavigationStories.DefaultStory} />
 
-### In DialogFullScreen
+### In Dialog
 
-`AnchorNavigation` can also be used as the content of `DialogFullScreen` component.
+`AnchorNavigation` can also be used as the content of `Dialog` component.
 
 <Canvas of={AnchorNavigationStories.InFullScreenDialogStory} />
