+++ src/components/anchor-navigation/components.test-pw.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState } from "react";
 
 import Textbox from "../textbox";
 import Button from "../button";
-import DialogFullScreen from "../dialog-full-screen";
+import Dialog from "../dialog";
 import {
   AnchorNavigation,
   AnchorSectionDivider,
@@ -38,7 +38,8 @@ export const InFullScreenDialog = () => {
   return (
     <>
       <Button onClick={() => setIsOpen(true)}>Open AnchorNavigation</Button>
-      <DialogFullScreen
+      <Dialog
+        fullscreen
         open={isOpen}
         onCancel={() => setIsOpen(false)}
         title="Title"
@@ -77,7 +78,7 @@ export const InFullScreenDialog = () => {
             <Content title="Fifth section" />
           </div>
         </AnchorNavigation>
-      </DialogFullScreen>
+      </Dialog>
     </>
   );
 };
