--- src/components/dialog-full-screen/dialog-full-screen.mdx
@@ -1,118 +1,0 @@
-import { Meta, ArgTypes, Canvas } from "@storybook/blocks";
-import TranslationKeysTable from "../../../.storybook/utils/translation-keys-table";
-
-import * as DialogFullScreenStories from "./dialog-full-screen.stories";
-
-<Meta of={DialogFullScreenStories} />
-
-# DialogFullScreen
-
-A full-width and full-height dialog overlaid on top of any page.
-
-## Contents
-
-- [Quick Start](#quick-start)
-- [Related Components](#related-components)
-- [Examples](#examples)
-- [Props](#props)
-- [Translation keys](#translation-keys)
-
-## Quick Start
-
-```javascript
-import DialogFullScreen from "carbon-react/lib/components/dialog-full-screen";
-```
-
-## Related Components
-
-- Simple task in context? [Try Dialog](../?path=/docs/dialog--docs).
-- Need to refer back to the underlying page? [Try Sidebar](../?path=/docs/sidebar--docs).
-
-## Examples
-
-### Default
-
-The call-to-action element should always be focused when the `DialogFullScreen` is closed. However, in some instances it may not receive focus
-due to specific browser design choices. The below example shows how to programmatically focus the call-to-action element when the `DialogFullScreen` is closed
-to ensure behaviour is consistent across all browsers.
-
-<Canvas of={DialogFullScreenStories.Default} />
-
-### Preventing focus from being restored when DialogFullScreen closes
-
-When the `restoreFocusOnClose` prop is `false`, focus will not be restored to the element that was focused before the `DialogFullScreen` was opened.
-Focus can instead be programmatically applied to another element if appropriate. 
-
-<Canvas of={DialogFullScreenStories.RestoreFocusOnCloseStory} />
-
-### With complex example
-
-If you want to use more than one group of `Tabs` remember to put `selectedTabId` prop in every `Tabs` component
-to have a possibility to manage active `Tabs` group
-
-<Canvas of={DialogFullScreenStories.WithComplexExample} />
-
-### With disableContentPadding
-
-<Canvas of={DialogFullScreenStories.WithDisableContentPadding} />
-
-### With header children
-
-<Canvas of={DialogFullScreenStories.WithHeaderChildren} />
-
-### With help
-
-<Canvas of={DialogFullScreenStories.WithHelp} />
-
-### With hideable header children
-
-<Canvas of={DialogFullScreenStories.WithHideableHeaderChildren} />
-
-### Overriding the first focused element
-
-By default, when a dialog is opened it will automatically focus the first element within its children that can be focussed.
-However, there are a couple of ways of overriding this default behaviour. The `focusFirstElement` prop accepts a reference
-to the element you wish to focus on open (an example of this can be seen by clicking the first button below). If the element
-you want to focus supports `autoFocus` then the you can override the default behaviour by using the `disableAutoFocus` prop
-and setting the `autoFocus` on the element you wish to be focused instead (click the second button to see an example).
-
-<Canvas of={DialogFullScreenStories.FocusingADifferentFirstElement} />
-
-### Allowing other elements to be focused
-
-Using the `focusableContainers` prop allows elements that are not DOM children of the Dialog to be reachable with the keyboard while the Dialog is open.
-This may occasionally be useful with things like Toasts where they persist on the page until dismissed by clicking a close icon.
-
-<Canvas of={DialogFullScreenStories.OtherFocusableContainers} />
-
-### Top modal override
-
-When multiple modals are open on a page Carbon manages the order internally so that the last one mounted in the DOM
-is treated as the top one and the focus trap is set on it. However, this can potentially
-create unexpected issues if the loading of the dialogs isn't synchronous. In order to guarantee that a given
-modal is treated as top set the `topModalOverride` prop. Please note that in the scenario where this prop is
-set on multiple modals, the last one mounted with it will be treated as top.
-
-<Canvas of={DialogFullScreenStories.TopModalOverride} />
-
-## Props
-
-### DialogFullScreen
-
-<ArgTypes of={DialogFullScreenStories} />
-
-## Translation keys
-
-The following keys are available to override the translations for this component by passing in a custom locale object to the
-[i18nProvider](../?path=/docs/documentation-i18n--docs).
-
-<TranslationKeysTable
-  translationData={[
-    {
-      name: "dialogFullScreen.ariaLabels.close",
-      description: "The text for close button aria-label attribute",
-      type: "func",
-      returnType: "string",
-    },
-  ]}
-/>
