+++ src/components/dialog-full-screen/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./dialog-full-screen.component";
-export type { DialogFullScreenProps } from "./dialog-full-screen.component";
