+++ src/components/dialog/dialog.test.tsx
@@ -10,6 +10,7 @@ import userEvent from "@testing-library/user-event";
 
 import CarbonProvider from "../carbon-provider";
 import Dialog, { DialogHandle, DialogProps } from ".";
+import Logger from "../../__internal__/utils/logger";
 
 beforeEach(() => {
   jest.useFakeTimers();
@@ -20,368 +21,489 @@ afterEach(() => {
   jest.useRealTimers();
 });
 
-test("dialog element has aria-modal attribute set to true when open", () => {
-  render(
-    <CarbonProvider>
-      <Dialog open title="My dialog" />
-    </CarbonProvider>,
-  );
-  expect(screen.getByRole("dialog")).toHaveAttribute("aria-modal", "true");
-});
+describe("Deprecation warnings", () => {
+  test("logs a deprecation warning when disableClose is used", () => {
+    const loggerSpy = jest.spyOn(Logger, "deprecate");
+    render(<Dialog fullscreen open disableClose />);
 
-test("title is displayed as a level 1 heading when title prop is passed", async () => {
-  render(<Dialog open title="My dialog" />);
+    expect(loggerSpy).toHaveBeenCalledWith(
+      "The disableClose prop in Dialog is deprecated and will soon be removed.",
+    );
+    expect(loggerSpy).toHaveBeenCalledTimes(1);
 
-  const dialog = screen.getByRole("dialog", { name: /My dialog/i });
-  const heading = within(dialog).getByRole("heading", {
-    level: 1,
-    name: /My Dialog/i,
+    loggerSpy.mockRestore();
   });
 
-  expect(heading).toBeVisible();
+  test("logs a deprecation warning when pagesStyling is used", async () => {
+    const loggerSpy = jest.spyOn(Logger, "deprecate");
+    render(<Dialog fullscreen pagesStyling open />);
+
+    expect(loggerSpy).toHaveBeenCalledWith(
+      "The pagesStyling prop in Dialog is deprecated and will soon be removed.",
+    );
+    expect(loggerSpy).toHaveBeenCalledTimes(1);
+
+    loggerSpy.mockRestore();
+  });
 });
 
-test("subtitle is displayed when subtitle prop is passed", async () => {
-  render(<Dialog open title="My dialog" subtitle="My subtitle" />);
+describe("Modal Dialog", () => {
+  test("passes className prop to the dialog element", () => {
+    render(<Dialog open title="My dialog" className="custom-class" />);
 
-  const dialog = screen.getByRole("dialog", { description: /My subtitle/i });
-
-  expect(dialog).toHaveTextContent("My subtitle");
-});
-
-test("custom title is displayed when title prop is a React element", () => {
-  render(<Dialog open title={<h1>Custom title</h1>} />);
-
-  const heading = screen.getByRole("heading", {
-    name: /Custom title/i,
-    level: 1,
+    const dialog = screen.getByTestId("modal");
+    expect(dialog).toHaveClass("custom-class");
   });
 
-  expect(heading).toBeVisible();
-});
-
-test("custom subtitle is displayed when subtitle prop is a React element", () => {
-  render(<Dialog open title="My dialog" subtitle={<h2>My subtitle</h2>} />);
-
-  const heading = screen.getByRole("heading", {
-    name: /My subtitle/i,
-    level: 2,
+  test("dialog element has aria-modal attribute set to true when open", () => {
+    render(
+      <CarbonProvider>
+        <Dialog open title="My dialog" />
+      </CarbonProvider>,
+    );
+    expect(screen.getByRole("dialog")).toHaveAttribute("aria-modal", "true");
   });
 
-  expect(heading).toBeVisible();
-});
+  test("title is displayed as a level 1 heading when title prop is passed", async () => {
+    render(<Dialog open title="My dialog" />);
 
-test("aria-label prop is used as dialog's accessible name, when passed and title isn't", () => {
-  render(<Dialog open aria-label="foobar" />);
-  expect(screen.getByRole("dialog")).toHaveAccessibleName("foobar");
-});
+    const dialog = screen.getByRole("dialog", { name: /My dialog/i });
+    const heading = within(dialog).getByRole("heading", {
+      level: 1,
+      name: /My Dialog/i,
+    });
 
-test("aria-labelledby prop is used as dialog's accessible name when passed", () => {
-  render(
-    <Dialog
-      open
-      title={<h1 id="title-id">Custom title</h1>}
-      aria-labelledby="title-id"
-    />,
-  );
-  expect(screen.getByRole("dialog")).toHaveAccessibleName("Custom title");
-});
+    expect(heading).toBeVisible();
+  });
 
-test("aria-describedby prop is used as dialog's accessible description when passed", () => {
-  render(
-    <Dialog
-      open
-      title="My dialog"
-      subtitle={<h2 id="subtitle-id">Custom subtitle</h2>}
-      aria-describedby="subtitle-id"
-    />,
-  );
-  expect(screen.getByRole("dialog")).toHaveAccessibleDescription(
-    "Custom subtitle",
-  );
-});
+  test("subtitle is displayed when subtitle prop is passed", async () => {
+    render(<Dialog open title="My dialog" subtitle="My subtitle" />);
 
-test("help icon is displayed when help prop is passed", () => {
-  render(<Dialog open title="My dialog" help="Help text" />);
+    const dialog = screen.getByRole("dialog", { description: /My subtitle/i });
 
-  const help = screen.getByLabelText("help");
+    expect(dialog).toHaveTextContent("My subtitle");
+  });
 
-  expect(help).toBeVisible();
-});
+  test("custom title is displayed when title prop is a React element", () => {
+    render(<Dialog open title={<h1>Custom title</h1>} />);
 
-test("close button is displayed when onCancel prop is passed", () => {
-  render(<Dialog open title="My dialog" onCancel={() => {}} />);
+    const heading = screen.getByRole("heading", {
+      name: /Custom title/i,
+      level: 1,
+    });
 
-  const closeButton = screen.getByRole("button", { name: /Close/i });
+    expect(heading).toBeVisible();
+  });
 
-  expect(closeButton).toBeVisible();
-});
+  test("custom subtitle is displayed when subtitle prop is a React element", () => {
+    render(<Dialog open title="My dialog" subtitle={<h2>My subtitle</h2>} />);
 
-test("close button is not rendered when showCloseIcon prop is false", () => {
-  render(<Dialog open onCancel={() => {}} showCloseIcon={false} />);
-  expect(
-    screen.queryByRole("button", { name: /Close/i }),
-  ).not.toBeInTheDocument();
-});
+    const heading = screen.getByRole("heading", {
+      name: /My subtitle/i,
+      level: 2,
+    });
 
-describe("closing behaviour", () => {
-  const MockApp = ({ onCancel }: Pick<DialogProps, "onCancel">) => {
-    const [open, setOpen] = React.useState(true);
-    return (
+    expect(heading).toBeVisible();
+  });
+
+  test("aria-label prop is used as dialog's accessible name, when passed and title isn't", () => {
+    render(<Dialog open aria-label="foobar" />);
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("foobar");
+  });
+
+  test("aria-labelledby prop is used as dialog's accessible name when passed", () => {
+    render(
       <Dialog
-        open={open}
+        open
+        title={<h1 id="title-id">Custom title</h1>}
+        aria-labelledby="title-id"
+      />,
+    );
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("Custom title");
+  });
+
+  test("aria-describedby prop is used as dialog's accessible description when passed", () => {
+    render(
+      <Dialog
+        open
         title="My dialog"
-        onCancel={(ev) => {
-          onCancel?.(ev);
-          setOpen(false);
+        subtitle={<h2 id="subtitle-id">Custom subtitle</h2>}
+        aria-describedby="subtitle-id"
+      />,
+    );
+    expect(screen.getByRole("dialog")).toHaveAccessibleDescription(
+      "Custom subtitle",
+    );
+  });
+
+  test("help icon is displayed when help prop is passed", () => {
+    render(<Dialog open title="My dialog" help="Help text" />);
+
+    const help = screen.getByLabelText("help");
+
+    expect(help).toBeVisible();
+  });
+
+  test("close button is displayed when onCancel prop is passed", () => {
+    render(<Dialog open title="My dialog" onCancel={() => {}} />);
+
+    const closeButton = screen.getByRole("button", { name: /Close/i });
+
+    expect(closeButton).toBeVisible();
+  });
+
+  test("close button is not rendered when showCloseIcon prop is false", () => {
+    render(<Dialog open onCancel={() => {}} showCloseIcon={false} />);
+    expect(
+      screen.queryByRole("button", { name: /Close/i }),
+    ).not.toBeInTheDocument();
+  });
+
+  describe("closing behaviour", () => {
+    const MockApp = ({ onCancel }: Pick<DialogProps, "onCancel">) => {
+      const [open, setOpen] = React.useState(true);
+      return (
+        <Dialog
+          open={open}
+          title="My dialog"
+          onCancel={(ev) => {
+            onCancel?.(ev);
+            setOpen(false);
+          }}
+        />
+      );
+    };
+
+    test("when close button is clicked, dialog closes and onCancel callback is executed", async () => {
+      const onCancel = jest.fn();
+      const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+
+      render(
+        <CarbonProvider>
+          <MockApp onCancel={onCancel} />
+        </CarbonProvider>,
+      );
+
+      const closeButton = screen.getByRole("button", { name: /Close/i });
+      await user.click(closeButton);
+      await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
+
+      expect(onCancel).toHaveBeenCalledTimes(1);
+    });
+
+    test("when Escape key is pressed, dialog closes and onCancel callback is executed", async () => {
+      const onCancel = jest.fn();
+      const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+      render(
+        <CarbonProvider>
+          <MockApp onCancel={onCancel} />
+        </CarbonProvider>,
+      );
+
+      await user.keyboard("{Escape}");
+      await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
+
+      expect(onCancel).toHaveBeenCalledTimes(1);
+    });
+
+    test("when close button is focused and the Enter key is pressed, dialog closes and onCancel callback is executed", async () => {
+      const onCancel = jest.fn();
+      const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+      render(
+        <CarbonProvider>
+          <MockApp onCancel={onCancel} />
+        </CarbonProvider>,
+      );
+
+      const closeButton = screen.getByRole("button", { name: /Close/i });
+      act(() => closeButton.focus());
+      await user.keyboard("{Enter}");
+      await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
+
+      expect(onCancel).toHaveBeenCalledTimes(1);
+    });
+
+    test("when close button is focused and the Space key is pressed, dialog closes and onCancel callback is executed", async () => {
+      const onCancel = jest.fn();
+      const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+      render(
+        <CarbonProvider>
+          <MockApp onCancel={onCancel} />
+        </CarbonProvider>,
+      );
+
+      const closeButton = screen.getByRole("button", { name: /Close/i });
+      act(() => closeButton.focus());
+      await user.keyboard("{ }");
+      await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
+
+      expect(onCancel).toHaveBeenCalledTimes(1);
+    });
+
+    test("when close button is focused and a non-Space/Enter key is pressed, dialog stays open and onCancel is not executed", async () => {
+      const onCancel = jest.fn();
+      const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+
+      render(
+        <CarbonProvider>
+          <MockApp onCancel={onCancel} />
+        </CarbonProvider>,
+      );
+
+      const closeButton = screen.getByRole("button", { name: /Close/i });
+      act(() => closeButton.focus());
+      await user.keyboard("{a}");
+
+      expect(onCancel).not.toHaveBeenCalled();
+    });
+  });
+
+  test("root container is refocused when the focus method of the component's ref handle is called", async () => {
+    const MockApp = () => {
+      const dialogHandle = React.useRef<DialogHandle>(null);
+      return (
+        <Dialog open title="My dialog" ref={dialogHandle}>
+          <button type="button" onClick={() => dialogHandle.current?.focus()}>
+            Refocus dialog
+          </button>
+        </Dialog>
+      );
+    };
+    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+    render(
+      <CarbonProvider>
+        <MockApp />
+      </CarbonProvider>,
+    );
+
+    const button = screen.getByRole("button", { name: /Refocus dialog/i });
+    await user.click(button);
+
+    expect(screen.getByRole("dialog")).toHaveFocus();
+  });
+
+  test("first focusable element is not focused when disableAutoFocus prop is passed", () => {
+    render(
+      <CarbonProvider>
+        <Dialog open title="My dialog" disableAutoFocus>
+          <button type="button">Focus me</button>
+        </Dialog>
+      </CarbonProvider>,
+    );
+
+    const button = screen.getByRole("button", { name: /Focus me/i });
+    expect(button).not.toHaveFocus();
+  });
+
+  test("height prop controls the dialog's height", () => {
+    render(<Dialog open title="My dialog" height="500px" />);
+    expect(screen.getByRole("dialog")).toHaveStyle({ height: "500px" });
+  });
+
+  test("maximum height of the dialog is 90% of the viewport height", () => {
+    render(<Dialog open title="My dialog" />);
+    expect(screen.getByRole("dialog")).toHaveStyle({
+      maxHeight: "90vh",
+    });
+  });
+
+  test("dialog element has correct data-* props", () => {
+    render(
+      <Dialog open title="My dialog" data-element="foo" data-role="bar" />,
+    );
+
+    const dialog = screen.getByRole("dialog", { name: /My dialog/i });
+
+    expect(dialog).toHaveAttribute("data-component", "dialog");
+    expect(dialog).toHaveAttribute("data-element", "foo");
+    expect(dialog).toHaveAttribute("data-role", "bar");
+  });
+
+  test("close button has correct data-* props, when the closeButtonDataProps prop is passed", () => {
+    render(
+      <Dialog
+        open
+        onCancel={() => {}}
+        closeButtonDataProps={{
+          "data-element": "foo",
+          "data-role": "bar",
         }}
-      />
-    );
-  };
-
-  test("when close button is clicked, dialog closes and onCancel callback is executed", async () => {
-    const onCancel = jest.fn();
-    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-
-    render(
-      <CarbonProvider>
-        <MockApp onCancel={onCancel} />
-      </CarbonProvider>,
+      />,
     );
 
     const closeButton = screen.getByRole("button", { name: /Close/i });
-    await user.click(closeButton);
-    await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
 
-    expect(onCancel).toHaveBeenCalledTimes(1);
+    expect(closeButton).toHaveAttribute("data-element", "foo");
+    expect(closeButton).toHaveAttribute("data-role", "bar");
   });
 
-  test("when Escape key is pressed, dialog closes and onCancel callback is executed", async () => {
-    const onCancel = jest.fn();
-    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-    render(
-      <CarbonProvider>
-        <MockApp onCancel={onCancel} />
-      </CarbonProvider>,
-    );
-
-    await user.keyboard("{Escape}");
-    await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
-
-    expect(onCancel).toHaveBeenCalledTimes(1);
+  test("renders with grey background when greyBackground prop is passed", () => {
+    render(<Dialog open title="My dialog" greyBackground />);
+    expect(screen.getByRole("dialog")).toHaveStyle({
+      backgroundColor: "var(--colorsUtilityMajor025)",
+    });
   });
 
-  test("when close button is focused and the Enter key is pressed, dialog closes and onCancel callback is executed", async () => {
-    const onCancel = jest.fn();
-    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-    render(
-      <CarbonProvider>
-        <MockApp onCancel={onCancel} />
-      </CarbonProvider>,
-    );
+  test("dialog is positioned correctly, when size prop is maximise", () => {
+    render(<Dialog open title="My dialog" size="maximise" />);
 
-    const closeButton = screen.getByRole("button", { name: /Close/i });
-    act(() => closeButton.focus());
-    await user.keyboard("{Enter}");
-    await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
-
-    expect(onCancel).toHaveBeenCalledTimes(1);
-  });
-
-  test("when close button is focused and the Space key is pressed, dialog closes and onCancel callback is executed", async () => {
-    const onCancel = jest.fn();
-    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-    render(
-      <CarbonProvider>
-        <MockApp onCancel={onCancel} />
-      </CarbonProvider>,
-    );
-
-    const closeButton = screen.getByRole("button", { name: /Close/i });
-    act(() => closeButton.focus());
-    await user.keyboard("{ }");
-    await waitForElementToBeRemoved(() => screen.queryByRole("dialog"));
-
-    expect(onCancel).toHaveBeenCalledTimes(1);
-  });
-
-  test("when close button is focused and a non-Space/Enter key is pressed, dialog stays open and onCancel is not executed", async () => {
-    const onCancel = jest.fn();
-    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-
-    render(
-      <CarbonProvider>
-        <MockApp onCancel={onCancel} />
-      </CarbonProvider>,
-    );
-
-    const closeButton = screen.getByRole("button", { name: /Close/i });
-    act(() => closeButton.focus());
-    await user.keyboard("{a}");
-
-    expect(onCancel).not.toHaveBeenCalled();
-  });
-});
-
-test("root container is refocused when the focus method of the component's ref handle is called", async () => {
-  const MockApp = () => {
-    const dialogHandle = React.useRef<DialogHandle>(null);
-    return (
-      <Dialog open title="My dialog" ref={dialogHandle}>
-        <button type="button" onClick={() => dialogHandle.current?.focus()}>
-          Refocus dialog
-        </button>
-      </Dialog>
-    );
-  };
-  const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-  render(
-    <CarbonProvider>
-      <MockApp />
-    </CarbonProvider>,
-  );
-
-  const button = screen.getByRole("button", { name: /Refocus dialog/i });
-  await user.click(button);
-
-  expect(screen.getByRole("dialog")).toHaveFocus();
-});
-
-test("first focusable element is not focused when disableAutoFocus prop is passed", () => {
-  render(
-    <CarbonProvider>
-      <Dialog open title="My dialog" disableAutoFocus>
-        <button type="button">Focus me</button>
-      </Dialog>
-    </CarbonProvider>,
-  );
-
-  const button = screen.getByRole("button", { name: /Focus me/i });
-  expect(button).not.toHaveFocus();
-});
-
-test("height prop controls the dialog's height", () => {
-  render(<Dialog open title="My dialog" height="500px" />);
-  expect(screen.getByRole("dialog")).toHaveStyle({ height: "500px" });
-});
-
-test("maximum height of the dialog is 90% of the viewport height", () => {
-  render(<Dialog open title="My dialog" />);
-  expect(screen.getByRole("dialog")).toHaveStyle({
-    maxHeight: "90vh",
-  });
-});
-
-test("dialog element has correct data-* props", () => {
-  render(<Dialog open title="My dialog" data-element="foo" data-role="bar" />);
-
-  const dialog = screen.getByRole("dialog", { name: /My dialog/i });
-
-  expect(dialog).toHaveAttribute("data-component", "dialog");
-  expect(dialog).toHaveAttribute("data-element", "foo");
-  expect(dialog).toHaveAttribute("data-role", "bar");
-});
-
-test("close button has correct data-* props, when the closeButtonDataProps prop is passed", () => {
-  render(
-    <Dialog
-      open
-      onCancel={() => {}}
-      closeButtonDataProps={{
-        "data-element": "foo",
-        "data-role": "bar",
-      }}
-    />,
-  );
-
-  const closeButton = screen.getByRole("button", { name: /Close/i });
-
-  expect(closeButton).toHaveAttribute("data-element", "foo");
-  expect(closeButton).toHaveAttribute("data-role", "bar");
-});
-
-test("renders with grey background when greyBackground prop is passed", () => {
-  render(<Dialog open title="My dialog" greyBackground />);
-  expect(screen.getByRole("dialog")).toHaveStyle({
-    backgroundColor: "var(--colorsUtilityMajor025)",
-  });
-});
-
-test("dialog is positioned correctly, when size prop is maximise", () => {
-  render(<Dialog open title="My dialog" size="maximise" />);
-
-  const dialog = screen.getByRole("dialog");
-  expect(dialog).toHaveStyle(`
+    const dialog = screen.getByRole("dialog");
+    expect(dialog).toHaveStyle(`
     height: calc(100% - var(--spacing400));
     width: calc(100% - var(--spacing400));
   `);
-});
+  });
 
-test("prevents content from overflowing", () => {
-  render(
-    <Dialog open title="My dialog">
-      Content
-    </Dialog>,
-  );
-  expect(screen.getByTestId("dialog-content")).toHaveStyle("overflow-y: auto");
-});
-
-test("no padding is rendered around dialog content, when zero padding is specified via contentPadding prop", () => {
-  render(
-    <Dialog open title="My dialog" contentPadding={{ p: 0 }}>
-      Inner content
-    </Dialog>,
-  );
-
-  const content = screen.getByTestId("dialog-content");
-
-  expect(content).toHaveStyle({ padding: "var(--spacing000)" });
-});
-
-test("background scroll remains disabled when returning to outer dialog after closing inner dialog", async () => {
-  const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-  const MockApp = () => {
-    const [open, setOpen] = React.useState(true);
-    return (
-      <Dialog open title="Outer dialog">
-        <Dialog
-          open={open}
-          topModalOverride
-          title="Inner dialog"
-          onCancel={() => setOpen(false)}
-        />
-      </Dialog>
+  test("prevents content from overflowing", () => {
+    render(
+      <Dialog open title="My dialog">
+        Content
+      </Dialog>,
     );
-  };
+    expect(screen.getByTestId("dialog-content")).toHaveStyle(
+      "overflow-y: auto",
+    );
+  });
 
-  render(
-    <CarbonProvider>
-      <MockApp />
-    </CarbonProvider>,
-  );
-  expect(screen.getByRole("dialog")).toHaveAccessibleName("Inner dialog");
+  test("no padding is rendered around dialog content, when zero padding is specified via contentPadding prop", () => {
+    render(
+      <Dialog open title="My dialog" contentPadding={{ p: 0 }}>
+        Inner content
+      </Dialog>,
+    );
 
-  await user.click(screen.getByRole("button", { name: /Close/i }));
+    const content = screen.getByTestId("dialog-content");
 
-  expect(screen.getByRole("dialog")).toHaveAccessibleName("Outer dialog");
-  expect(document.body).toHaveStyle("overflow: hidden");
+    expect(content).toHaveStyle({ padding: "var(--spacing000)" });
+  });
+
+  test("background scroll remains disabled when returning to outer dialog after closing inner dialog", async () => {
+    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
+    const MockApp = () => {
+      const [open, setOpen] = React.useState(true);
+      return (
+        <Dialog open title="Outer dialog">
+          <Dialog
+            open={open}
+            topModalOverride
+            title="Inner dialog"
+            onCancel={() => setOpen(false)}
+          />
+        </Dialog>
+      );
+    };
+
+    render(
+      <CarbonProvider>
+        <MockApp />
+      </CarbonProvider>,
+    );
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("Inner dialog");
+
+    await user.click(screen.getByRole("button", { name: /Close/i }));
+
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("Outer dialog");
+    expect(document.body).toHaveStyle("overflow: hidden");
+  });
+
+  test("should render with the highlight when 'ai' is passed in as the `highlightVariant`", () => {
+    render(
+      <Dialog open title="My dialog" highlightVariant="ai">
+        Inner content
+      </Dialog>,
+    );
+
+    const highlightElement = screen.getByRole("dialog");
+    expect(highlightElement).toHaveStyleRule(
+      "background",
+      `linear-gradient( 90deg, #00d639 0%, #00d6de 40%, #9d60ff 90% )`,
+      {
+        modifier: "::before",
+      },
+    );
+
+    expect(highlightElement).toBeVisible();
+  });
+
+  test("no padding is rendered around dialog content, when zero padding is specified via disableContentPadding prop", () => {
+    render(
+      <Dialog open title="My dialog" disableContentPadding>
+        Inner content
+      </Dialog>,
+    );
+
+    const content = screen.getByTestId("dialog-content");
+
+    expect(content).toHaveStyle({ padding: "0px" });
+  });
 });
 
-test("should render with the highlight when 'ai' is passed in as the `highlightVariant`", () => {
-  render(
-    <Dialog open title="My dialog" highlightVariant="ai">
-      Inner content
-    </Dialog>,
-  );
+describe("Fullscreen Dialog", () => {
+  test("renders the fullscreen dialog correctly", () => {
+    render(
+      <Dialog fullscreen open>
+        <button type="button" data-role="first-child">
+          first child
+        </button>
+        <div data-role="second-child">second child</div>
+        <p data-role="third-child">third child</p>
+      </Dialog>,
+    );
 
-  const highlightElement = screen.getByRole("dialog");
-  expect(highlightElement).toHaveStyleRule(
-    "background",
-    "linear-gradient(90deg,#00D639 0%,#00D6DE 40%,#9D60FF 90%)",
-    {
-      modifier: "::before",
-    },
-  );
+    expect(screen.getByTestId("first-child")).toBeVisible();
+    expect(screen.getByTestId("second-child")).toBeVisible();
+    expect(screen.getByTestId("third-child")).toBeVisible();
+  });
 
-  expect(highlightElement).toBeVisible();
+  // test here for coverage only
+  test("padding is removed from the content when the `contentPadding` prop is passed", () => {
+    render(
+      <Dialog fullscreen open contentPadding={{ p: 0 }}>
+        <div>test content</div>
+      </Dialog>,
+    );
+
+    const content = screen.getByTestId("dialog-content");
+    expect(content).toHaveStyleRule("padding", "var(--spacing000)");
+  });
+
+  test("when the `title` prop is a string, this value is set as the dialog's accessible name", () => {
+    render(<Dialog fullscreen open title="Test" />);
+
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("Test");
+  });
+
+  test("when the `title` and `subtitle` props are provided, the subtitle is set as the dialog's accessible description", () => {
+    render(<Dialog fullscreen open title="Title" subtitle="Subtitle" />);
+
+    expect(screen.getByRole("dialog")).toHaveAccessibleDescription("Subtitle");
+  });
+
+  test("when the `title` prop is a React element and the `aria-labelledby` prop is provided, it is passed to the dialog container", () => {
+    render(
+      <>
+        <h2 id="test-id">custom title</h2>
+        <Dialog
+          fullscreen
+          open
+          title={<h2>test</h2>}
+          aria-labelledby="test-id"
+        />
+      </>,
+    );
+
+    expect(screen.getByRole("dialog")).toHaveAccessibleName("custom title");
+  });
+
+  test("no padding is rendered around dialog content, when zero padding is specified via disableContentPadding prop", () => {
+    render(
+      <Dialog open fullscreen title="My dialog" disableContentPadding>
+        Inner content
+      </Dialog>,
+    );
+
+    const content = screen.getByTestId("dialog-content");
+
+    expect(content).toHaveStyle({ padding: "0px" });
+  });
 });
