+++ src/components/form/form.stories.tsx
@@ -13,7 +13,6 @@ import Textarea from "../textarea";
 import Dialog from "../dialog";
 import DateInput, { DateChangeEvent } from "../date";
 import { Select, MultiSelect, Option } from "../select";
-import DialogFullScreen from "../dialog-full-screen";
 import { RadioButton, RadioButtonGroup } from "../radio-button";
 import { Checkbox } from "../checkbox";
 import Hr from "../hr";
@@ -337,7 +336,7 @@ export const WithButtonsAlignedToTheLeft: Story = (args: FormProps) => (
 );
 WithButtonsAlignedToTheLeft.storyName = "With Buttons Aligned to the Left";
 
-export const WithRequiredFields: Story = (args: FormProps) => (
+export const WithBothOptionalOrRequired: Story = (args: FormProps) => (
   <Box m={1}>
     <RequiredFieldsIndicator mb={2}>
       <Typography variant="b">Fill in all fields marked with</Typography>
@@ -351,13 +350,7 @@ export const WithRequiredFields: Story = (args: FormProps) => (
         </Button>
       }
     >
-      <Textbox label="Textbox" />
       <Textbox label="Textbox" required />
-      <Select name="simple-optional" id="simple-optional" label="Simple Select">
-        <Option text="Amber" value="1" />
-        <Option text="Black" value="2" />
-        <Option text="Blue" value="3" />
-      </Select>
       <Select
         name="simple-required"
         id="simple-required"
@@ -387,7 +380,12 @@ export const WithRequiredFields: Story = (args: FormProps) => (
         <Option text="Black" value="2" />
         <Option text="Blue" value="3" />
       </MultiSelect>
-      <RadioButtonGroup name="radio group optional" legend="RadioGroup">
+      <RadioButtonGroup
+        name="radio group optional"
+        legend="RadioGroup"
+        value="group-1-input-1"
+        onChange={() => "RADIO CHANGE"}
+      >
         <RadioButton
           id="group-1-input-1"
           value="group-1-input-1"
@@ -404,6 +402,8 @@ export const WithRequiredFields: Story = (args: FormProps) => (
       <RadioButtonGroup
         name="radio group required"
         legend="RadioGroup"
+        value="group-2-input-2"
+        onChange={() => "RADIO CHANGE"}
         required
       >
         <RadioButton
@@ -419,12 +419,12 @@ export const WithRequiredFields: Story = (args: FormProps) => (
           labelWidth={10}
         />
       </RadioButtonGroup>
-      <Checkbox name="checkbox" label="Checkbox" />
       <Checkbox name="checkbox" label="Checkbox" required />
     </Form>
   </Box>
 );
-WithRequiredFields.parameters = {
+WithBothOptionalOrRequired.storyName = "WithBothOptionalOrRequired";
+WithBothOptionalOrRequired.parameters = {
   themeProvider: { chromatic: { theme: "sage" } },
 };
 
@@ -563,7 +563,8 @@ export const InDialogFullScreen = () => {
   return (
     <>
       <Button onClick={() => setIsOpen(true)}>Open Preview</Button>
-      <DialogFullScreen
+      <Dialog
+        fullscreen
         open={isOpen}
         onCancel={() => setIsOpen(false)}
         title="Form in Dialog"
@@ -579,7 +580,7 @@ export const InDialogFullScreen = () => {
             <Textbox label="Textbox" />
           </Form>
         </Box>
-      </DialogFullScreen>
+      </Dialog>
     </>
   );
 };
@@ -591,7 +592,8 @@ export const InDialogFullScreenWithStickyFooter = () => {
   return (
     <>
       <Button onClick={() => setIsOpen(true)}>Open Preview</Button>
-      <DialogFullScreen
+      <Dialog
+        fullscreen
         open={isOpen}
         onCancel={() => setIsOpen(false)}
         title="Form in Dialog"
@@ -632,7 +634,7 @@ export const InDialogFullScreenWithStickyFooter = () => {
             <Textbox key={`textbox-${index + 1}`} label="Textbox" />
           ))}
         </Form>
-      </DialogFullScreen>
+      </Dialog>
     </>
   );
 };
@@ -683,6 +685,8 @@ export const FormAlignmentExample: Story = (args: FormProps) => {
         legendWidth={10}
         legendSpacing={2}
         legendAlign="right"
+        value="group-1-input-1"
+        onChange={() => "RADIO CHANGE"}
       >
         <RadioButton
           id="group-1-input-1"
@@ -707,7 +711,13 @@ export const FormAlignmentExample: Story = (args: FormProps) => {
           setDate(ev.target.value.formattedValue)
         }
       />
-      <RadioButtonGroup name="nolegend" legend="Legend above" ml="10%">
+      <RadioButtonGroup
+        name="nolegend"
+        legend="Legend above"
+        ml="10%"
+        value={"group-2-input-2"}
+        onChange={() => "RADIO CHANGE"}
+      >
         <RadioButton
           id="group-2-input-1"
           value="group-2-input-1"
@@ -729,6 +739,8 @@ export const FormAlignmentExample: Story = (args: FormProps) => {
         labelInline
         labelWidth={10}
         inputWidth={30}
+        value=""
+        onChange={() => {}}
       />
       <Checkbox name="checkbox1" label="Checkbox 1" ml="10%" />
       <Checkbox name="checkbox2" label="Checkbox 2" ml="10%" />
@@ -754,6 +766,8 @@ export const FormAlignmentExample: Story = (args: FormProps) => {
         labelWidth={10}
         labelSpacing={2}
         mb={4}
+        checked
+        onChange={() => "SWITCH CHANGE"}
       />
       <Textbox
         key="input-five"
