+++ src/components/pages/pages.mdx
@@ -31,7 +31,7 @@ import Pages, { Page } from "carbon-react/lib/components/pages";
 
 <Canvas of={PagesStories.WithInitialPageIndex} />
 
-## Inside DialogFullScreen
+## Inside Dialog
 
 <Canvas of={PagesStories.InsideDialogFullScreen} />
