+++ src/components/select/simple-select/simple-select.pw.tsx
@@ -25,7 +25,7 @@ import {
   helpIcon,
   tooltipPreview,
 } from "../../../../playwright/components";
-import { alertDialogPreview } from "../../../../playwright/components/dialog";
+import { dialogWithRole } from "../../../../playwright/components/dialog";
 import { loader } from "../../../../playwright/components/loader";
 import {
   dropdownButton,
@@ -1283,7 +1283,7 @@ test.describe("When nested inside of a Dialog component", () => {
 
     await selectText(page).click();
     const inputElement = commonDataElementInputPreview(page);
-    const dialogElement = alertDialogPreview(page);
+    const dialogElement = dialogWithRole(page, "dialog");
     await inputElement.press("Escape");
     await expect(selectList(page)).not.toBeVisible();
     await expect(dialogElement).toBeVisible();
@@ -1298,7 +1298,7 @@ test.describe("When nested inside of a Dialog component", () => {
     await mount(<SimpleSelectNestedInDialog />);
 
     await selectText(page).click();
-    await alertDialogPreview(page).click();
+    await dialogWithRole(page, "dialog").click();
     await expect(selectList(page)).not.toBeVisible();
     await expect(commonDataElementInputPreview(page)).not.toBeFocused();
   });
