+++ src/components/sidebar/sidebar.mdx
@@ -37,7 +37,6 @@ import Sidebar from "carbon-react/lib/components/sidebar";
 ## Related Components
 
 - Simple task in context? [Try Dialog](../?path=/docs/dialog--docs).
-- Complex task that needs more space? [Try Dialog Full Screen](../?path=/docs/dialog-full-screen--docs).
 
 ## Examples
 
@@ -52,7 +51,7 @@ to ensure behaviour is consistent across all browsers.
 ### Preventing focus from being restored when Sidebar closes
 
 When the `restoreFocusOnClose` prop is `false`, focus will not be restored to the element that was focused before the `Sidebar` was opened.
-Focus can instead be programmatically applied to another element if appropriate. 
+Focus can instead be programmatically applied to another element if appropriate.
 
 <Canvas of={SidebarStories.RestoreFocusOnCloseStory} />
