+++ src/locales/fr-ca.ts
@@ -53,11 +53,6 @@ const frCA: Partial<Locale> = {
       close: () => "Fermer",
     },
   },
-  dialogFullScreen: {
-    ariaLabels: {
-      close: () => "Fermer",
-    },
-  },
   dismissibleBox: {
     ariaLabels: {
       close: () => "Fermer",
